requirements.txt
Django==1.6.2
MySQL-python==1.2.5
Pillow==2.4.0
South==0.8.4
django-admin-sortable2==0.2.7
django-resized==0.2.4
django-tinymce==1.5.2
wsgiref==0.1.2
